iced_dialog::dialog

Struct Dialog

pub struct Dialog<'a, Message, Theme = Theme, Renderer = Renderer>
where
    Renderer: 'a + Renderer,
    Theme: 'a + Catalog,
{ /* private fields */ }

A message dialog.

Only the content is required, buttons and the title are optional.

The sizing strategy used depends on whether you add any buttons: if you don’t, the dialog will be sized to fit the content similarly to a container. If you do add buttons, max_width & max_height (or width & height when set to a fixed pixel value) are used. If these aren’t set, DEFAULT_MAX_WIDTH and/or DEFAULT_MAX_HEIGHT are used.

Implementations

impl<'a, Message, Theme, Renderer> Dialog<'a, Message, Theme, Renderer>
where Renderer: 'a + Renderer + Renderer, Theme: 'a + Catalog, Message: 'a + Clone,

pub fn new( is_open: bool, base: impl Into<Element<'a, Message, Theme, Renderer>>, content: impl Into<Element<'a, Message, Theme, Renderer>>, ) -> Self

Creates a new Dialog with the given base and dialog content.

pub fn with_buttons( is_open: bool, base: impl Into<Element<'a, Message, Theme, Renderer>>, content: impl Into<Element<'a, Message, Theme, Renderer>>, buttons: Vec<Element<'a, Message, Theme, Renderer>>, ) -> Self

Creates a new Dialog with the given base, dialog content and buttons.

pub fn title(self, title: impl IntoFragment<'a>) -> Self

Sets the Dialog’s title.

pub fn on_press(self, on_press: Message) -> Self
where Message: Clone,

Sets the message that will be produced when the Dialog’s backdrop is pressed.

pub fn on_press_with(self, on_press: impl Fn() -> Message + 'a) -> Self

Sets the message that will be produced when the Dialog’s backdrop is pressed.

This is analogous to Dialog::on_press, but using a closure to produce the message.

pub fn on_press_maybe(self, on_press: Option<Message>) -> Self
where Message: Clone,

Sets the message that will be produced when the Dialog’s backdrop is pressed, if Some.

pub fn width(self, width: impl Into<Length>) -> Self

Sets the Dialog’s width.

pub fn height(self, height: impl Into<Length>) -> Self

Sets the Dialog’s height.

pub fn max_width(self, max_width: impl Into<Pixels>) -> Self

Sets the Dialog’s maximum width.

pub fn max_height(self, max_height: impl Into<Pixels>) -> Self

Sets the Dialog’s maximum height.

pub fn align_left(self) -> Self

Aligns the Dialog to the left.

pub fn align_right(self) -> Self

Aligns the Dialog to the right.

pub fn align_top(self) -> Self

Aligns the Dialog to the top.

pub fn align_bottom(self) -> Self

Aligns the Dialog to the bottom.

pub fn align_x(self, alignment: impl Into<Horizontal>) -> Self

Sets the Dialog’s alignment for the horizontal axis.

Dialogs are horizontally centered by default.

pub fn align_y(self, alignment: impl Into<Vertical>) -> Self

Sets the Dialog’s alignment for the vertical axis.

Dialogs are vertically centered by default.

pub fn padding_inner(self, padding: impl Into<Padding>) -> Self

Sets the Dialog’s inner padding.

pub fn padding_outer(self, padding: impl Into<Padding>) -> Self

Sets the Dialog’s outer padding (sometimes called “margin”).

pub fn spacing(self, spacing: impl Into<Pixels>) -> Self

Sets the Dialog’s spacing.

pub fn align_buttons(self, align: impl Into<Vertical>) -> Self

Sets the vertical alignment of the Dialog’s buttons.

pub fn font(self, font: impl Into<Renderer::Font>) -> Self

Sets the Font of the Dialog’s title.

pub fn push_button( self, button: impl Into<Element<'a, Message, Theme, Renderer>>, ) -> Self

Adds a button to the Dialog.

pub fn push_button_maybe( self, button: Option<impl Into<Element<'a, Message, Theme, Renderer>>>, ) -> Self

Adds a button to the Dialog, if Some.

pub fn extend_buttons( self, buttons: impl IntoIterator<Item = Element<'a, Message, Theme, Renderer>>, ) -> Self

Extends the Dialog with the given buttons.

pub fn backdrop(self, color: impl Into<Color>) -> Self
where <Theme as Catalog>::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the backdrop color of the Dialog.

pub fn style(self, style: impl Fn(&Theme) -> Style + 'a) -> Self
where <Theme as Catalog>::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the Dialog.

pub fn title_style(self, style: impl Fn(&Theme) -> Style + 'a) -> Self
where <Theme as Catalog>::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the Dialog’s title.

pub fn container_style(self, style: impl Fn(&Theme) -> Style + 'a) -> Self
where <Theme as Catalog>::Class<'a>: From<StyleFn<'a, Theme>>,

Sets the style of the Dialog’s container.

pub fn class(self, class: impl Into<<Theme as Catalog>::Class<'a>>) -> Self

Sets the style class of the Dialog.

pub fn title_class( self, class: impl Into<<Theme as Catalog>::Class<'a>>, ) -> Self

Sets the style class of the Dialog’s title.

pub fn container_class( self, class: impl Into<<Theme as Catalog>::Class<'a>>, ) -> Self

Sets the style class of the Dialog’s container.